split cipher.rs
Previously the `cipher` mod was very large. This commit splits it into
three modules:

* `certificate` for the certificate, certified key, and root store
  code.
* `verifier` for the client/server verifier code.
* `cipher` for the remaining ciphersuite/protocol code.

References are updated throughout.
