+++ src/client.rs
@@ -13,7 +13,7 @@ use rustls::{
     KeyLogFile, ProtocolVersion, SignatureScheme, SupportedProtocolVersion,
 };
 
-use crate::cipher::{rustls_certified_key, rustls_server_cert_verifier};
+use crate::certificate::rustls_certified_key;
 use crate::connection::{rustls_connection, Connection};
 use crate::crypto_provider::{rustls_crypto_provider, rustls_hpke};
 use crate::error::rustls_result::{InvalidParameter, NullParameter};
@@ -21,6 +21,7 @@ use crate::error::{self, map_error, rustls_result};
 use crate::keylog::{rustls_keylog_log_callback, rustls_keylog_will_log_callback, CallbackKeyLog};
 use crate::rslice::NulByte;
 use crate::rslice::{rustls_slice_bytes, rustls_slice_slice_bytes, rustls_str};
+use crate::verifier::rustls_server_cert_verifier;
 use crate::{
     arc_castable, box_castable, crypto_provider, ffi_panic_boundary, free_arc, free_box,
     set_arc_mut_ptr, set_boxed_mut_ptr, to_boxed_mut_ptr, try_box_from_ptr, try_clone_arc,
